requirements.txt
pandas==2.1.4
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0